refactor: lookup 서비스를 패키지로 묶기
