+++ src/main/java/com/techfork/post/application/query/lookup/PostKeywordLookupService.java
@@ -1,4 +1,4 @@
-package com.techfork.post.application.query;
+package com.techfork.post.application.query.lookup;
 
 import com.techfork.post.domain.PostKeyword;
 import com.techfork.post.infrastructure.PostKeywordRepository;
