+++ src/test/java/com/techfork/activity/bookmark/application/command/BookmarkCommandServiceTest.java
@@ -5,7 +5,7 @@ import com.techfork.activity.bookmark.domain.BookmarkErrorCode;
 import com.techfork.activity.bookmark.infrastructure.BookmarkRepository;
 import com.techfork.post.domain.Post;
 import com.techfork.post.domain.exception.PostErrorCode;
-import com.techfork.post.application.query.PostLookupService;
+import com.techfork.post.application.query.lookup.PostLookupService;
 import com.techfork.domain.source.entity.TechBlog;
 import com.techfork.domain.useraccount.entity.User;
 import com.techfork.domain.useraccount.exception.UserErrorCode;
