+++ src/test/java/com/techfork/activity/readpost/application/command/ReadPostCommandServiceTest.java
@@ -7,7 +7,7 @@ import com.techfork.activity.readpost.infrastructure.ReadPostRepository;
 import com.techfork.post.application.command.PostViewCountCommandService;
 import com.techfork.post.domain.Post;
 import com.techfork.post.domain.exception.PostErrorCode;
-import com.techfork.post.application.query.PostLookupService;
+import com.techfork.post.application.query.lookup.PostLookupService;
 import com.techfork.domain.source.entity.TechBlog;
 import com.techfork.domain.useraccount.entity.User;
 import com.techfork.domain.useraccount.exception.UserErrorCode;
